% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{add_header}
\alias{add_header}
\title{Add a header to the subgroup table}
\usage{
add_header(
  data,
  group1_lbl = c("", ""),
  group2_lbl = c("", ""),
  or_lbl = c("Odds ratio", "[95\% CI]"),
  p_lbl = "p"
)
}
\arguments{
\item{data}{a data frame generated by \code{subgroup_table_binomial()}.}

\item{group1_lbl}{a character vector containing a label for treatment group 1.
Each row of the vector will be translated into a separate line in the
resulting data set.}

\item{group2_lbl}{a character vector containing a label for treatment group 2.}

\item{or_lbl}{a character vector containing a label for the odds ratio.}

\item{p_lbl}{a character vector containing a label for the p value.}
}
\value{
A tibble ready for input into \code{forestplot::forestplot()}.
}
\description{
This function adds a header to the table generated by \code{subgroup_table_binomial()}.
}
\examples{
library(dplyr)
data(knee, package="catdata")

knee \%>\% mutate(Th = factor(Th),
                Sex = factor(Sex),
                R4_improved = factor(if_else(R4 < R1, 1, 0))) ->
  knee
formula_overall = "R4_improved ~ Th + R1"
knee \%>\%
  subgroup_table_binomial(data = .,
                          formula_overall,
                          formulas = list(
                            Sex = "R4_improved ~ Th*Sex + R1"
                          ),
                          formulas_subgroups = list(
                            Sex = formula_overall
                          ),
                          group = "Th",
                          outcome = "R4_improved",
                          outcome_positive = 1)  \%>\%
  add_header(., group1_lbl = c("Improvement in", "placebo group"),
             group2_lbl = c("Improvement in", "treatment group")) ->
  res
}
\seealso{
\link{subgroup_table_binomial} generates the table for which this header is intended.
}
