% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rd}
\alias{rd}
\title{Round a decimal number to two digits}
\usage{
rd(x)
}
\arguments{
\item{x}{a number.}
}
\value{
A string of the input number rounded to two digits.
}
\description{
Round a decimal number to two digits
}
