% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rd_p}
\alias{rd_p}
\title{Round a p-value to three digits}
\usage{
rd_p(x)
}
\arguments{
\item{x}{a number.}
}
\value{
A string of the input number rounded to three digits or string
containing "<0.001" if the number is less than 0.001.
}
\description{
Round a p-value to three digits
}
