% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rd_pct}
\alias{rd_pct}
\title{Round a percentage to two digits}
\usage{
rd_pct(x)
}
\arguments{
\item{x}{a number.}
}
\value{
A string of the input number as percentage rounded to two digits
(i.e. 0 digits in terms of percentage).
}
\description{
Round a percentage to two digits
}
